Move Shockwave engine classes to its own sub-folder to have better separation between game logic and engine/helper classes
