+++ GTFO_VR/Core/PlayerBehaviours/BodyHaptics/Shockwave/Engine/HapticGroupInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace GTFO_VR.Core.PlayerBehaviours.BodyHaptics.Shockwave
+﻿namespace GTFO_VR.Core.PlayerBehaviours.BodyHaptics.Shockwave.Engine
 {
     public struct HapticGroupInfo
     {
