+++ GTFO_VR/Core/PlayerBehaviours/BodyHaptics/Shockwave/Engine/HapticIndex.cs
@@ -1,4 +1,4 @@
-﻿namespace GTFO_VR.Core.PlayerBehaviours.BodyHaptics.Shockwave
+﻿namespace GTFO_VR.Core.PlayerBehaviours.BodyHaptics.Shockwave.Engine
 {
     public class HapticIndex
     {
