+++ GTFO_VR/GTFO_VR.csproj
@@ -46,12 +46,12 @@
     <Compile Include="Core\PlayerBehaviours\BodyHaptics\ElevatorSequenceAgent.cs" />
     <Compile Include="Core\PlayerBehaviours\BodyHaptics\ElevatorSequenceIntegrator.cs" />
     <Compile Include="Core\PlayerBehaviours\BodyHaptics\ElevatorState.cs" />
-    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\HapticGroupInfo.cs" />
-    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\HapticGroupPattern.cs" />
-    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\HapticIndex.cs" />
-    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\HapticIndexPattern.cs" />
+    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\Engine\HapticGroupInfo.cs" />
+    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\Engine\HapticGroupPattern.cs" />
+    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\Engine\HapticIndex.cs" />
+    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\Engine\HapticIndexPattern.cs" />
+    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\Engine\ShockwaveEngine.cs" />
     <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\ShockwaveElevatorSequence.cs" />
-    <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\ShockwaveEngine.cs" />
     <Compile Include="Core\PlayerBehaviours\BodyHaptics\Shockwave\ShockwaveIntegration.cs" />
     <Compile Include="Core\PlayerBehaviours\MovementVignette.cs" />
     <Compile Include="Core\UI\Terminal\Pointer\TerminalPointer.cs" />
